Added a second position-sensitive response from Nagisa to Streaming-chan
Added a response from Nagisa to Streaming-chan's "girl on girl" question. This one is much easier to test because there's a 1/3 chance she'll ask this on Nagisa's first strip in a two-opponent game. (Streaming-chan starts with a low viewer count.) Note that this response currently only happens at Nagisa's stage zero, but I will expand it to all stages in the future.

To test: Streaming-chan on left, Nagisa on right. Make Nagisa lose in debug mode, and restart if Streaming-chan doesn't ask about girl-on-girl. Then try again with Nagisa on the left and Streaming-chan on the right.

Also made the following two changes to the response Nagisa gained in the previous commit:
* No need for the notSaidMarker self-checks when using alsoPlayingSayingMarker checks: Streaming-chan will only ask these questions once. I still marker the lines so that she won't reply again in the following phase, though.
* I got un-lazy and found the right cases instead of using "global" everywhere. For this, they are must_strip, stripping, and stripped for when Nagisa is on the left, and only the latter two for when she's on the right.
I also applied these changes to the new response.