Let a random player matching a `<condition>` be bound to a custom dialogue variable.
The variable can also be used in `<test>`s to further constrain the case and who the variable can be bound to (all combinations of possible variable bindings from conditions will be tested). One of the combinations of variable bindings will be saved in the case object.

The State constructor is modified to take a Case reference, which is used to access both the custom variable bindings and the sets of tags to be added and removed.

The `<condition>` element gets a new attribute `role`, which can be used to match the current target, all players other than the subject (whether the target or not), or all other players other than the target and the subject (like alsoPlaying). The count also gets a default of "one or more".

I've also added a `character` attribute (to match a character ID) and a `stage` attribute (corresponding to `targetStage` and `alsoPlayingStage` depending on the `role`). The idea is to add all the `target`/`alsoPlaying` attributes to allow multiple sets of `alsoPlaying` conditions and reduce the number of attributes, but there's some work to be done with the priority calculations.

In its simplest use case, one can do

`<condition var="random">`
`<state img="0-happy.png">Hello, ~random~!</state>`