Nugi-chan epilogue adjustment
Nugi-chan now mildly chastises the player for viewing her epilogue via the Gallery instead of at the end of a real game.

I did this with a marker. I set a marker on Nugi-chan's defeated lines. If that marker is not set, Nugi-chan knows you reached her reward via the back entrance.