Collectibles: implement counter-based collectibles
Collectibles can now have an optional <counter> subelement, specifying a counter
value that must be reached before a collectible can be unlocked.

Collectible counters work in exactly the same way as marker counters,
and support roughly the same set of operations (increment, decrement, set, and unlock).

To support this, <state> elements now have a `collectible-value` attribute in
addition to a `collectible` attribute, to specify values for incrementing,
decrementing, or setting collectible counters:

 - Numbers prefixed with '+' will be interpreted as 'increment counter' commands.
   (i.e. '+5' = increment by 5, '+1' = increment by 1)
 - Numbers prefixed with '-' will be interpreted as 'decrement counter' commands.
   (i.e. '-5' = decrement by 5, '-1' = decrement by 1)
 - Unprefixed numbers will be interpreted as 'set counter' commands.
   (i.e. '5' = set to 5, '1' = set to 1).

Note that counter values are always >= 0.
If a `collectible-value` attribute is not specified, the default action is to
unlock the specified collectible, setting any attached counter to its maximum value.