Fix version string comparison on Firefox
tl;dr Firefox does not support _named capture groups_ for some dumb reason, despite it being part of the ES2018 standard for over a year now, and despite it being the _only_ major browser to lack support.

There are many things I want to do first thing in the morning. Debugging cross-browser regex compatibility is not one of them.