Wide pose handling and adjustable opponent layering
This primarily adds a `<z-layer>` option to `meta.xml` to allow creators to
explicitly override the usual left-right opponent layering rules if absolutely necessary.

This also adds a `<dialogue-layering>` option with possible values of `over` or `under`,
to specify if the player's dialogue box should layer over or under player poses, respectively.
(TBH, this could also be made into a State attribute.)
The default value for this option is `over`.

Finally, this also tweaks the `.opponent-image` CSS so that poses are always
horizontally centered, even if they overflow their container area.