Added Beta's new 6-stripped image to Nugi-chan
Beta submitted a new pose for Nugi-chan as part of Nugi-chan's Single-Stage Pose Promotion. (See her channel for details) He's the only one that's entered so far, so the emoji he suggested is looking pretty likely to be in the top six of the popular vote!

I've included this new pose on five new conditional lines for Nugi-chan as she strips: first exposed, only female, 4+ consecutive losses, SPH, or street background.

I also added a couple of new targeted lines.

625 -> 633 unique lines