Added Nugi-chan collectible, adjusted Nagisa's epilogue gender requirements
Added a collectible to Nugi-chan using the new CE; another one to come shortly

Removed the male version of the "Night Out" epilogue. The UI in this epilogue indicates that the player is wearing a bra, and it was originally only intended as a female-only epilogue. As the character select screen can now better message the requirements for epilogues, I have removed what was effectively a duplicate (but it had one fewer screen than the female version).

I've also adjusted the "Starting Out" epilogue to be male only. Back in December, I added a basic penetration animation, as I forgot that was enabled for female players. Sorry!

Nagisa's epilogues now:
- "Starting Out" (male only, was any)
- "Girls' Night Out" (female only, had male variant)
- "A Friendly Chat" (still for any)