Add player variables that provide more information about how a player is doing compared to others:
`~player.place~`: Current placement (1-5) based on layers remaining. If two players are tied, the lower number is returned, as per usual. When a player is out, their place is locked, so the first player to masturbate in a five player game stays at place 5 and don't become tied for fourth when the next player is eliminated.

`~player.revplace~`: Current placement in reverse order, mainly so that you can check who is last or second last regardless of the number of participants. Also the lower number in case of a tie. The first player out will have this locked at 1 and so on.

`~player.lead~`: Difference between `player`'s number of remaining layers and that of the best of the rest, so you can gloat when you're in a really big lead. If `player` is not in the lead, this will be negative, reflecting the number of layers behind the leader.

`~player.trail~`: Difference between the lowest number of layers remaining among the other *remaining* players and `player`'s number of layers. Negative if not actually in last place (among the remaining players).

`~player.diff(other)~` Difference between the number of layers of `player` and `other`, default being `self`.

As an example, you can pinpoint the current leader (if any) by using `<condition var="leader"/><test expr="~leader.lead~" op="&gt;" value="0"/>`

This commit moves the code that looks up the player a variable refers to into a separate function.