Make the saying and sayingMarker attributes of <condition> work as expected with updateVolatileBehaviour.
In the process, simplify how updateVolatileBehaviour works and make *SaidMarker and *NotSaidMarker conditions react to changes in the current phase.

There are no changes to how updateBehaviour(), updateAllBehaviours(), updateVolatileBehaviour(), or updateAllVolatileBehaviours() are called, but instead of collecting potential volatile cases in a separate array and processing so-called volatile conditions in a separate function, they are checked together with the ordinary conditions and dependencies are recorded as we go. Then on the volatile passes, only cases with higher priorities are searched. Instead of incrementing and decrementing a lock counter, we search the recorded dependencies, which is fairly efficient.

This means that characters in the initial update pass of updateAllBehaviours() can react to markers set by previous players (to avoid them reacting to markers set by players to the right in the previous phase, a flag is set).

When a *SaidMarker condition reacts to a marker that wasn't previously set, it works as *SayingMarker and records a dependency.

*NotSaidMarker conditions now also allow the same kinds of expressions as *SaidMarker, but the result is negated.

The way saidMarker, saying, etc. of the <condition> element are handled is a bit complicated because you can in theory have a `<condition count="0" saying="Hello"/>`, that is, you want to make sure that nobody else is saying "Hello". After determining that nobody is saying "Hello", we can't allow the state of another player to change, because someone might then say "Hello", violating the condition. 99 times out of 100, you'd check if at least one character is saying something, however.