Extended Formatting: <hr> element support and data-character attributes
A `data-character` attribute is now added to the wrapper spans used for class specifiers.
This allows for CSS rules of the form:

```css
.my-dialogue-class[data-character="my-character"] {
    /* Styling... */
}
```

This should help prevent issues due to using a shared CSS class namespace for all characters.

Unescaping of the <hr> (horizontal rule) element is now performed.
<hr> elements can be used as a divider between logically separate elements of dialogue,
such as when separating dialogue for joint characters.