Make fn matching part of dialogue expansion regexp non-greedy but match-everything.
By making it non-greedy, later parts of the regexp will be tried first, so we don't have to exclude any characters.

As it was, the regexp would consume everything until the last `~` in `dialogue`, so more than one variable expansion in the same string wouldn't work.

Also use console.error() on expansion errors to make them easier to see, and leave tildes in place if marker or collectible expansions are used without specifying a marker or collectible.