Nugi-chan lines
I'm playing around with the preview of version 5.0 of the Character Editor. The "Play Once" feature is pretty awesome. Click the "..." button after the line and tick the box to ensure that the line you chose will not play more than once per game. No markers required!

I played with this a little and added some random targeted lines to Nugi-chan as I did.

820->835 unique lines

If you would like Nugi-chan to say something to your character, leave a comment in her channel naming your favorite fictional maid (that isn't Nugi-chan). Open to all.