﻿Monika 5.0.0: 'Poiesis'
In philosophy, _poiesis_ (from Ancient Greek: ποίησις) is "the activity in which a person brings something into being that did not exist before."

A variant of the term is also the root of the word 'poetry'.

Changelog:

 - Updated and refactored all custom Monika JS code to hopefully improve stability
 - Near-completely rewrote Effect implementation code
 - Changed random glitches to be less intrusive
 - Changed scaling of glitch chances throughout game
 - Fixed Sayonika interaction
 - Fixed glitch forfeit mode

Note: I've performed basic testing for this update, but some lingering instability may well remain.