Redefine card and card area sizes in terms of rem in order to make them more exact numbers and all so that all the cards have the same proportions (3:4), adjusting sizes slightly. Also move player cards and clothes to the bottom middle of the spaces between the main button and the left and right screen edges when using the minimal UI.
This simplifies the CSS a bit, resizing the cards instead of the containers.

The animations no longer require the special #hidden-large-card image; we can instead clone the visible deck card.

The prompt-exchange class is now set on the game-screen element and the animated card is full size when exchanging cards.