Make Clothing a class and change genericName/formalName (with genericName as the default) to just name and generic (with name as default).
Includes suggested generics for the human player's clothing and changes the position of the belt to 'waist'.

make_xml.py is changed such that if the first field of a `clothing` line starts with an lowercase letter, it interprets the line as name,generic,type,etc.; otherwise as the old format.