Add start_offline.exe and start_offline.sh
start_offline.exe is a fully pre-packaged version of the NodeJS http-server
package, built for Windows x86. Just run it and it'll automatically start
up a local webserver and open your browser to the offline version.

start_offline.sh is a script that runs http-server via npx, for users
on Linux and OSX.

You'll need to install NodeJS and NPM first, before using it though:
https://nodejs.org/en/download/