Set z-index on .screen to make sure that negative z-indices on opponents make them end up underneath the screen div.
Eliminate .dialogue-bubble-row (to not have bubbles layered under that either), .image-row (for consistency and simplicity), and the distinction between .opponent-image-area and .selection-image-area, using the former everywhere and making it 25% wide (the width doesn't matter anyway).

Eliminating .dialogue-bubble-row Requires changes to .stats-box a bit.

Explicitly set pointer-events: auto on input elements just to be sure.

Place suggestions in the background. This requires pointer-events: none also on #main-select-screen .opponent-image-area.