Added nicknames to Nugi-chan
Long has Nugi-chan called everybody ~name~-sama, except in targeted lines. With the Nicknames feature, she can now use honorifics properly! This includes -chan for characters with that built in (like herself and Streaming-chan) -tan for Zone-tan, and -hime for princesses.

This update affected a lot of lines of dialogue where the honorifics were previously baked in, so please report any errors if you see them.