Add after_masturbating situation that plays during the exchange phase after a character has finished masturbating (falling back to finished_masturbating).
I moved the handling of fallback situations into `updateBehaviour()` to avoid complicating the code in `advanceTurn()` that deals with eliminated players. If `tags` is a *nested* array, it's interpreted as a list of sets of situations to test in order. It should work because there happens to be no calls to `updateAllBehaviours()` passing a flat array as the `other_tags` argument - it's all single situations or nested arrays.

Closes #17.