Enabled Nugi-chan's costume dialogue
The game engine has a bug that prevents it from testing variable expressions that contain a hyphen, must like the one in Nugi-chan's costume name.

For now, I've removed the hyphen from the costume name, but we'll need to fix up the engine because this could cause other new lines with advanced cases to unexpectedly not work in the future.

Please feel free to test Nugi-chan in her new costume before the event to find dialogue bugs to report, especially if she mentions wearing her regular outfit.