Change inInterval() to return null if the argument can't be parsed (Closes #29)
If the argument contained a "-", it would return an object (with max and/or min set to NaN) even if what was before and after the "-" wasn't numeric, but null if there was no "-" and the argument couldn't be parsed as a number.

Now it will return null unless what's before and after the "-" is either nothing or a number.

This should avoid problems when testing a variable against a string containing a "-".