Restored Nugi-chan's costume name hyphen
ReformCopyright fixed the bug that prevented reskins with hyphens from working properly, so I've restored the one in Nugi-chan's file (hopefully) before anyone targeted it. I will keep a lookout for any characters targeting the non-hyphenated version this week and will advise targeters if I see it.

If your character already targeted Nugi-chan's new costume, an easy way to fix this would be to open up your character's dialogue file and do a find-replace for the variable name: "nugichan_bikini" to "nugi-chan_bikini".