More summer dialogue for Nugi-chan
924 -> 943 unique lines

Nugi-chan now has more to say to Sayaka, Sheena, and Monika. Over the last few days, I've added new lines for all of the characters with summer alts.

Something I could expand on more is the acknowlegement of the excellent Tiki Bar background. But I didn't find any lines that were particularly incongruous with it in these rounds of testing.