Hanako: copy 9-embarrassed.gif -> 9-embarrassed.png
CE v5.1 might not properly handle changing image extensions across stages.
This is a temporary workaround until that's fixed.

Despite having a .png extension, the image is still a proper animated GIF.

This relies on browsers ignoring the reported Content-Type for this
particular image and doing file type detection on their own, so it's kinda
iffy and sketchy.