Implement alternative conditions by copying cases
This is most useful together with !7.

Simple example:

```xml
<case tag="hand" notSaidMarker="hello">
  <condition character="ruby"/>
  <alternative>
     <condition character="weiss"/>
  </alternative>
  <alternative>
     <condition character="yang"/>
  <alternative>
  <state>...</state>
</case>
```
The above would then play if Ruby is present along with either Weiss or Yang (or both), without having to duplicate all conditions and lines in the file (they are just duplicated in memory).

Since oneShotId is copied as well, a one-shot case with alternative conditions will still only play once, not once per alternative.