Support multiple images per `<state>`, specified by stage.
Requires putting the text in a `<text>` element.

Example syntax is `<state img="#-normal.png"><img stage="1-2 4">#-special.png</img><img stage="3-6">#-badass.png</img><text>How do you like this?</text></state>`

If more than one `<img>` is valid for a stage, one is picked randomly. If none is, the `img` attribute is used as the default.

I chose to pass the current stage as an argument to the Case and State constructors and pick an image when each State is instantiated in order to affect other code less.