Add dealing_cards trigger, which will play when cards are dealt except when someone is masturbating.
I envision this to be used only in special situations for now. For example, when Revy attacks Meia, if Meia's slot is to the far right and Revy's to the far left, Revy will appear in two places for quite a bit, unless the AI speed is "Very fast" (more accurately called "Instantaneous"). Conversely, if Meia is to the left, Revy will momentarily disappear. This gives an opportunity for her to get back to her place before the cards are dealt.

It could also be used to finish stripping in case the `must_strip` - `stripping` - `stripped` sequence isn't enough, without having to stand half-stripped until their turn. I do think that this kind of extended stripping sequence should be avoided, though.

It can look weird if only some characters' behaviour is updated while the rest stay still, but I think we can get away with it if it's to avoid characters appearing double. In this case, I expect characters to say nothing and just change their poses, but just in case, I made it so that `dealing_cards` lines are skipped if the deal animation and AI speeds are so high that the lines wouldn't be seen.