Avoid calling updateGameVisual(), repeating animations, when no state change was made (closes #61)
Because updateGameVisual() is always called after Opponent.prototype.commitBehaviourUpdate(), the former is now called from the latter, but only when there's a new chosenState to commit.

Duplicate dialogue transcript entries were also created when the cards are dealt. To avoid that, saveAllTranscriptEntries() is now called from commitAllBehaviourUpdates(), which records which players had new states to commit (meaning it can be called from advanceTurn() after the extra reaction round).