Added even more Testing Tables!
Yep, there are a lot! But there's no cost to adding more, and I think we're at a place where we need a few more because:

* There are 21 characters on testing!
* Each of the masquerade characters should be on a table with each other masquerade character at least once. We're doing our best to encourage them to target each other.
* Each character should be on at least two tables.

The major change here is that I've added several tables to the start. These are the 21 characters with no repetitions to prevent any character from being buried more than 6 tables down. (Sorry, Rainbow Dash!)

After the masquerade contest ends, we can cut these down to a smaller number as testing empties out again.