Selection: add default table filling
This adds a mechanism for filling in the selection screen slots on
startup with a random set of characters with specific highlight or
status values.

This can be controlled via the `<default-fill>` configuration option.

If this is set to a value other than "none", it will enable default
table filling using characters with the specified highlight or status
attribute.

For example, specifying `new` for `<default-fill>` will cause the game
to randomly select new characters on startup, specifying `testing` will
cause the game to pick random Testing characters, etc.