Highlighting: add badges for new characters and for alt costumes
Thanks to CorkyTheCactus, we now have shiny new icons for denoting
characters that are new and that have alternate costumes, similiarly to
how epilogues are denoted on the select screens.

To accomodate this, the layout of badges on the group select screen and
on opponent selection cards has been changed to a dynamic flexbox-based
layout, instead of using absolute positioning. This should ensure that
badges stack properly without having to play around with positioning
values.