Added two collectibles to Nugi-chan
As Nugi-chan is at 1059 lines, she is eligible for up to three collectibles: one for existing (which must be for defeating her or winning the game against her). one for having between 0-600 lines, and one for having 600-1200 lines. One day when Nugi-chan gets more than 1200 lines, she'll be eligible for another collectible.

The excellent images for these collectibles were provided by UnderscorM3!