Updated testing tables again
Panty, Chazz, and Saki 265918 are on the main roster now, so they've given up their testing table positions to other characters.

Due to fewer characters of testing, I merged four tables. (For example, if removing a character made one table of A/B/C and one table of B/C/D, I moved D to the first table and removed the second table.)