Deactivated the re-sort modal for this year
The re-sort modal has done an amazing job of getting more players to vote in the first two re-sort polls! We had more votes in the first poll this re-sort than in all four polls last re-sort!

For the third poll, we're going to see how those results compare to people who have more of a finger on the pulse of development.