Changed "Targeted most by selected" to "Talks to selected"
The "Targeted most by selected" option is an effective character selection sorting method to get more interactive games. And using the word "target" makes a lot of sense when describing internally who a line is to be spoken to.

However, using the word "targeted" in the UI of the character selection screen is jargon and implies that the characters you're sorting for won't just speak to her or him, they'll help focus fire him down somehow, as though the game will be easier by matching one opponent with people they're weak against.

Several suggestions were made to replace this phrase in the past year, and Vis suggested "Talks to selected" back in September (#56). This strikes the right balance of describing the sort without implying that the characters listed will necessarily reply.