Adds several CE recipes
MisterSolitaire's prize for the SPNatI 4th Anniversary fanart competition was for me to improve one of his character's behaviour files and create line cases very specific to her for MisterSolitaire to write lines for.

However, as MisterSolitaire is already a proficient character author, he has instead asked me to write custom cases that *anybody* could use and to find a way to share and store these as Character Editor recipes, as per issue #68. Specifically, he asked me to focus on cases that cater to unusual strip orders, as we often get reports based on those.

I co-ordinated with spnati_edit, and as of version 5.5 of the Character Editor, the CE now supports recipes found in the `recipes` subfolder of your offline version. And, extra awesomely, spnati_edit has added a feature that shows checkmarks when you use each recipe!

Try these out and let me know if you find any bugs.