Add backwards compatibility for the legacy persist-marker flag
We already do a full traversal of all cases/states during character load, to
calculate outbound targeted line counts. This adds an additional check during
that traversal for states that have `persist-marker="true"` set, and if so
adds their associated markers to the opponent's `persistentMarkers` list.

Naturally, migrating to the new forward-declaration system when possible
is still a priority, so we can get rid of this backwards-compatibility code quickly.