Game over modal wording adjustments + tips
This changes the wording on the ending epilogue modal as suggested in #71.
The text for losing and for winning without any available epilogues is mostly
the same; in neither case do we mention epilogues.
(The win requirement for epilogues is already described in the FAQ.)

The text for winning with available epilogues does mention "extra" content
by necessity, because it's a prompt.

This also adds randomly-selected tips that are displayed in between the
header text and the list of epilogues (if applicable).
These are displayed in all cases.

Resolves #71.