Calls out a couple of Situations for Mister Clean
April 1 approaches, and it may pay to write a few targeted lines to the game's existing April Fool's Day characters in preparation for their return. This update calls out a pair of Situations for Mister Clean to encourage this.

As Mister Clean's behaviour.xml was already in v5.2.1, the diff here is minimal.