Add numbered variables beside the one bound randomly.
Example:

```xml
<condition gender="male" count="2-3" var="man"/>
<test expr="~man.lead~" value="1" />
```

This will, if there are 2-3 males and one of them is in the lead, depending on the number of males, also randomly define `~man2~` and perhaps `~man3~` (but you can only use `~man2~` since you don't know the exact number).

As before, the game will feed all combinations of (primary) variable bindings that fulfil the condition elements to the variable tests. All possible permutations of the additional numbered variables will not be tried, but as long as the count is 2, that's not a problem.