Trigger fullscreen on double-click, but avoid doing so when double-clicking an input element or any bordered element (which most input elements are anyway).
This doesn't work perfectly, as the first click sometimes causes another element to appear and catch the second click, preventing bubbling from being prevented but not preventing the non-delegated dblclick event.

This requires disabling double-tap to zoom. Firefox seems to honor touch-action:manipulation a bit arbitrarily, but we can at least set touch-action:none on input elements to remove the delayed reaction.

pointer-events is set to none on the banner just so we don't have to exclude it from the selector of elements to exclude from fullscreen activation.