Fix epilogue "Previous" button UI error
EpiloguePlayer.hasPreviousDirectives() was incorrectly assuming
that all epilogues start at sceneIndex=0, directiveIndex=1, though
this doesn't hold for epilogues that have multiple directives before
their first pause / wait / delay directive.

This bug was causing the "Previous" button to remain incorrectly enabled
when returning to the start of these epilogues.

Technically, this bug isn't fixed in the case of hot reloads, because
honestly, it's such a trivial issue that it's not worth adding
additional code to determine the correct first directive index for
hot-reloaded epilogues. This issue only affects UI, and most devs
should be able to live with it.

But it's fixed in the user-facing case, at least.