Video player style epilogue buttons.
Puts glyphicon-based buttons in a bar along the bottom. A separate div along the left edge allows going backwards.

The home (previously "Reset?") button is always enabled and visible.

The bar is hidden when moved away from, except for the first three seconds at the start, at the end, and when debug mode is active.

Backward and forward are disabled until the epilogue is loaded.

The buttons are scaled differently than everything else - only based on windows height.

Also fixes the bug that the keyboard handler for the epilogue screen isn't detached on leaving said screen.