Prevent players from entering before all characters are loaded
Adds a small loading meter that ensures players don't progress
to the title screen before all character metadata is loaded.

Trying to enter the select screens or the Gallery before all
characters are loaded tends to lead to issues, such as empty
gallery views and empty selection screens.

It's easier to prevent players from progressing to the title
screen entirely, than to make sure that players can't move
to the gallery and selection screens separately from the
title screen before everything is ready.