Load and execute marker operations within epilogues
`<marker>` elements can now be used as children of <epilogue>
elements, and the marker operations they describe will be
executed whenever the epilogue is played.

By default, epilogue marker operations only run when epilogues are
played at the conclusion of a game.

This can be controlled using the `when` attribute, which is only
valid for `<marker>` elements that are children of `<epilogue>`s.

There are three possible values for `when`:
 - `game` (default): Execute marker operation only at the
   end of a game.

 - `gallery`: Execute marker operation only when playing epilogues
   from the gallery.

 - `always`: Execute marker operation whenever this epilogue is
   played from both the gallery and after games.