Feedback: Indicate when feedback is disabled
Adds a warning message when leaving feedback for characters
that have feedback disabled.

This should be a softer alternative to a1d24e4e74, that avoids
the issues associated with completely disabling feedback for
characters whose creators may have forgotten to enable feedback
reporting, while making it clear why feedback may not be relayed
to Discord for those characters.

# Conflicts:
#	js/spniCore.js