spniBehaviour: specifically check customPriority for NaN
Fixes a bug introduceed by ba046290f6, where `priority="0"`
was misinterpreted as an undefined/invalid customPriority value.

Probably the best way I can explain this is:
```js
(NaN || undefined) === (0 || undefined) === undefined
```

Therefore, we need to explicitly check for `isNaN(this.customPriority)`
in the `Case` constructor and _only_ then do we set it to undefined.