Selection: distinguish epilogues with unmet requirements
Adds `Opponent.prototype.getAllEpilogueStatus` for summarizing
epilogue unlock status, and a function `getOpponentEpilogueStatus`
in `spniDisplay.js` that uses it to identify if any epilogues
can be unlocked for a character, given the current player gender
and table setup.

This is used in two places:
 - `OpponentDetailsDisplay` now uses this in its main `update`
   method to figure out whether a character's epilogue status
   should be displayed as "Available", "Conditionally Available",
   gender-locked, character-locked, etc.

 - `OpponentSelectionCard` now uses this to add a `.missing-reqs`
   class to the epilogue badges for characters with epilogues
   that cannot be unlocked at the moment (for example, due to
   alsoPlaying requirements).

At the moment, the `.missing-reqs` class simply applies a CSS
greyscale filter effect onto the epilogue badge it's attached
to, and also drops its opacity slightly (to 0.4).