Let players access the title screen, but not the select or gallery screens, until all metadata has been loaded.
Because people hate waiting without something to do in the meantime.

What we might want to do is to prevent entry to the title screen until the config and localStorage have been loaded.

I would like to let the player tag button be visible from the start but that messes up the layout.