Incremented make_xml.py to version 2.2
As per my last commit to this file, I believe that the era of make_xml has entirely passed and it no longer has purpose for us. However, I note that updates were made to it as recently as eight months ago.

I've updated the version stamping for the same reason as other times: to help us better understand what state make_xml was in when it generated a file we're assisting a contributor with. I don't think we'll have anyone come to us for help with a version 2.2 file, but people dig up old tools and old guides, so it's not impossible.