Vectorize as circles and lines instead, vectorize the any/all icons as well, add nominal size attribute, and actually replace the PNGs in the game.
The main reason for doing this is that it was suggested that the epilogue badges include gender icons to indicate when you need to switch genders to unlock them.

Note that in addition to looking cleaner, the SVG versions use a consistent stroke width (10% of the canvas size), which the PNGs didn't; the circles where slightly thinner.

This also moves the alt texts in the Epilogue gallery to the title attributes so that they might be seen.