Replace calls to `.find()` with `.children()` or an anchored selector (one that begins with ">"), so that jQuery doesn't have to search through all subelements looking for the <behaviour> element, for example.
This should speed up loading, shouldn't it?

NB: The documentation doesn't say that a selector that starts with ">" only matches starting from the current element, or even that it's valid, but that appears to be the case. Perhaps I need to change such queries into a chain of `.children()` calls too.

There were also some instances of overcomplicated code that checked if the result of a `$()` call was `undefined` which is never the case.