Small fix for Moon
A couple of Moon's lines re-used her "stripped" poses as versatile images for not-just-stripped situations. Two of Moon's new stripped images now show which clothing was removed and are no longer usable for this purpose.

Re-using "stripped" poses is usually a good idea, but showing what clothing was removed is a better idea.