Replaced Moon's selection screen pose to show her enthusiasm
Moon's calm.png image was not particularly calm previously, so it was used on some lines where she was expressing a lot of enthusiasm, notably on the character selection screen.

I looked over the lines where she used this line with an exclamation mark and have reassigned some poses accordingly.