Whitelisted event characters in config.xml for offline play
config.xml is used in the offline and rehosted editions in the game. Commit !1731 reclassified April Fool's Day characters + GSC as "event", and this category was not specifically whitelisted in config.xml, thereby hiding them for offline and rehosted play.

As with offline and incomplete characters, event characters should be part of the extended roster available to people offline. However, characters with the new "duplicate" tag continue to be intentionally hidden by default.