Adds Komi-san to the Seeking Sponsorship testing table
It's my understanding that the following four characters are currently seeking sponsorship:

* Erufuda: https://old.reddit.com/r/spnati/comments/fb0rc2/would_you_like_fries_with_that_elf_erufuda/
* Mary: https://old.reddit.com/r/spnati/comments/ggf2ai/mary_calvin_seeks_sponsorship/
* Shantae: https://old.reddit.com/r/spnati/comments/gn3arp/shantae_sponsorship_thread/
* Komi-san: https://old.reddit.com/r/spnati/comments/grdi4n/komisan_is_ready_to_make_some_new_friends_komisan/

As a reminder, if a character gets more than five sponsors, the creator can choose which five they wish to use. This is especially merciful if you believe what's asked of one or more of the sponsorships is too onerous.