Avoid orphaned card sprites if the player clicks too quickly and starts the next round before the card exchanging animation has finished
by aborting the animation on clicking Exchange.

Also, use the `$clonedCard` reference to remove the card sprite; no need to set an ID. That just risks problems because another sprite with the same ID could be created, and a search for an ID only returns a single element.

Add back the "shown-card" class removed in 7120173fec4d77841 because I thought it was only used for CSS and forgot it was also used by `stopCardAnimations()`.