Restored Mulan's collectibles
Mulan's Lucky Cricket collectible is now awarded via a hidden line. It is now for winning against her, as per the description, not just for her losing.

I have not had opportunity to exhaustively test the Fireworks collectible. As such, it is currently marked as hidden and will again be visible to people who had already unlocked it. I will test this thoroughly in a future update.