Continued Summer Event theming
Invited four more ladies in swimwear to the top of the roster for the duration of the event. These are characters that are not specifically featured and so will not be prefilled. One is offline, so she will only appear there during your offline games. This is the closest we've had to an official Swimsuit Edition! Of course, all types of summer attire are welcome for the event.

Later in the event, we will change the background from Beach to Tiki Bar to help our beachgoers get out of the sun.