Merula saved in CE v5.9
No content changes, just saved with new syntax. The next commit will contain fixes for Merula, which will be seen more readily in the diff.

This update appears to change tags, but that's only because this character's behaviour.xml and meta.xml were out of sync. Be sure to upload both with every update, and editor.xml too if you're not 1000% sure if it's changed or not.